+++ packages/orq-rc/RELEASES.md
@@ -779,3 +779,13 @@ Based on:
 - [typescript v3.2.0-rc.38] packages/orq-rc
 ### Releases
 - [NPM v3.2.0-rc.38] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.38 - packages/orq-rc
+
+## 2025-03-11 11:10:54
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.514.1 (2.546.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.2.0-rc.39] packages/orq-rc
+### Releases
+- [NPM v3.2.0-rc.39] https://www.npmjs.com/package/@orq-ai/node/v/3.2.0-rc.39 - packages/orq-rc